+++ ezCurator.py
@@ -1,60 +1,87 @@
-import requests, bs4, time, datetime, sys
+#!/usr/bin/env python3
+
+import re, sys, language_check 
+from beem.comment import Comment
 from beem.discussions import Query, Discussions_by_hot
 
-def getTime():
-    return datetime.datetime.fromtimestamp(
-            time.time()).strftime('%H:%M')
+def grammarCheck(baseText):
+    # use regex to remone HTML and Markdown syntax
+    text = syntaxStrip(baseText)
+    # TODO: allow user to select locale
+    tool = language_check.LanguageTool('en-US')
+    errors = tool.check(text)
+    # count sentences to check against errors
+    sent = text.split('.')
+    errorPerc = (len(errors)/words(text))*100
+    return round(errorPerc, 2)
 
-def getDate():
-    return datetime.datetime.fromtimestamp(
-            time.time()).strftime('%Y-%m-%d')
-
-def webCrawler(uri):
-    req = requests.get('https://busy.org/' + str(uri))
-    return bs4.BeautifulSoup(req.text, 'html.parser')
-
-def pub(postSoup):
-    return str(postSoup.select('.StoryFull__header__text__date span')[0])
-
-def body(postSoup):
-    return postSoup.select('.Body--full')
+def syntaxStrip(baseText):
+    return re.sub(r'(<.*>)|\#+|(\].*\))|\[|\*+|\>', '', baseText)
 
 def words(content):
-    return len(str(content).split())
+    # remove any HTML and Markdown
+    text = syntaxStrip(content)
+    return len(str(text).split())
 
-def published(uri):
-    return pub(webCrawler(uri))[6:-10]
-
-def content(uri):
-    return body(webCrawler(uri))
-
-def htmlOut(uri, uriShow):
+def htmlOut(uri, uriShow, content, published, body):
+    # Image to replace if none found in post
+    notFound = "https://media.giphy.com/media/8L0Pky6C83SzkzU55a/giphy.gif"
     # Output data with HTML
-    if words(content(uri)) >= 1000:
+    if words(content) >= 1000:
+        images = []
+        imgReg = r"http\S*.(jpg|png|jpeg|gif)"
+        
+        # get grammar check
+        errorScore = grammarCheck(body)
+
+        for text in content.split():
+            if ".jpg" in text or ".png" in text:
+                try:
+                    image = re.search(imgReg, text, re.IGNORECASE)
+                    images.append(str(image.group()))
+                except AttributeError:
+                    images.append(notFound)
+
+        try:
+            topImage = images[0]
+        except IndexError:
+            topImage = notFound
+
         curated.write('\n<div class="card mb-3">\n<div class="card-body">')
-        curated.write('\n<p><a href="https://steemit.com/' + uri + '">' + uriShow + 
-                '</a></p>')
-        curated.write('\n<p>Contains about ' + str(words(content(uri))) + ' words.</p>')
-        curated.write('\n<p>Published ' + str(published(uri)) + 'prior to ' + 
-                str(getTime()) + ' on ' + str(getDate()) + '</p>')
+        curated.write('\n<img class="card-img-top" style="max-width:500px" src="'+ 
+                topImage +'" alt="Card image cap">')
+        curated.write('\n<p><strong><a href="https://steemit.com/' + uri + '">' + uriShow + 
+                '</a></strong></p>')
+        curated.write('\n<p>Contains about ' + str(words(content)) + 
+                ' words with about  ' + str(errorScore) + 
+                '% of the text containing grammatical errors.</p>')
+        curated.write('\n<p>Published on ' + str(published)[:-15] + '</p>')
         curated.write('\n</div>\n</div>\n')
 
 def getPosts():
     # grab Steem URI
-    q = Query(limit=20, tag=str(sys.argv[1]))
-    for blog in Discussions_by_hot(q):
-        post = str(blog)
-        splitPost = post.split()
-        uriBase = splitPost[1]
+    q = Query(limit=50, tag=str(sys.argv[1]))
+    for entry in Discussions_by_hot(q):
+        # Get uri
+        postUri = str(entry)
+        splitUri = postUri.split()
+        uriBase = splitUri[1]
         uri = uriBase[:-1]
 
+        # grab post data from the blockchain
+        post = Comment(uri)
+        body = post['body']
+        published = post['created']
+
+        # cut uri display for uniformity on output
         if len(uriBase[:-1]) > 50:
             uriShow = str(uriBase[:47]) + '...'
         else:
             uriLen = len(uriBase[:-1])
             uriShow = uriBase[:-1] + '&nbsp;'*(50 - uriLen)
+
         # write HTML card data
-        htmlOut(uri, uriShow)
+        htmlOut(uri, uriShow, body, published, body)
 
 
 # open file to start HTML document
@@ -67,7 +94,7 @@ with open('top.html', 'r') as top:
             for topLine in top:
                 curated.write(topLine)
 
-            curated.write('<p>Below are the post with more than 1,000 words in the ' 
+            curated.write('<p style="text-align:center">Below are the post with more than 1,000 words in the ' 
                     + sys.argv[1] + ' "hot" feed.</p>\n')
             # Get Steem data
             getPosts()