+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Header images to cards
+- The [Beem](https://github.com/holgern/beem/tree/master/beem) Comment module to read post content
+- The language-check python package for grammar checking
+- Percent of possible grammar errors to HTML output
+
+### Removed
+- Beautiful Soup 4 web scraping
+- All functions dealing with web scraping
 
 ## [0.1.0](https://gitlab.com/jrswab/curator-helper/commit/6395f15beb6098a35f80f05a05f930522294ed5c) - 2018-08-27
 ### Added